feat: renames webpush app and only registers service worker when user is authenticated
