+++ apps/config/settings.py
@@ -95,7 +95,7 @@ LOCAL_APPS = (
     "apps.news",
     "apps.room",
     "apps.transaction",
-    "apps.web_push",
+    "apps.webpush",
 )
 
 INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + LOCAL_APPS
