+++ apps/templates/core/base.html
@@ -10,7 +10,10 @@
     <link rel="icon" href="{% static 'images/favicon.ico' %}">
 
     {% load_pwa_meta_data %}
-    {% load_serviceworker %}
+
+    {% if not request.user.is_anonymous %}
+        {% load_serviceworker %}
+    {% endif %}
 
     <!-- Bootstrap -->
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet"
