--- apps/web_push/models/__init__.py
@@ -1,1 +1,0 @@
-from apps.web_push.models.web_push_information import WebPushInformation  # noqa: F401
