+++ apps/webpush/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class WebPushConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "apps.web_push"
+    name = "apps.webpush"
