+++ apps/webpush/models/__init__.py
@@ -1,0 +1,1 @@
+from apps.webpush.models.web_push_information import WebPushInformation  # noqa: F401
